refactor: Extract animation settings into dedicated store
- Move animation level management from useSoundStore to new useAnimationsStore
- Add AnimationsInitializer component for proper store initialization
- Update imports and usage across affected components
- Maintain existing AsyncStorage patterns and error handling
