+++ frontend/app/components/initializers/StoreInitializer.tsx
@@ -9,6 +9,7 @@ import { useL2Store } from "@/app/stores/useL2Store";
 import { useTransactionsStore } from "@/app/stores/useTransactionsStore";
 import { useAchievementsStore } from "@/app/stores/useAchievementsStore";
 import { useSoundStore, useSound } from "@/app/stores/useSoundStore";
+import { useAnimationsStore } from "@/app/stores/useAnimationsStore";
 import { useTutorialStore } from "@/app/stores/useTutorialStore";
 import { useUpgradesStore } from "@/app/stores/useUpgradesStore";
 import { useTransactionPauseStore } from "@/app/stores/useTransactionPauseStore";
@@ -261,12 +262,23 @@ const TransactionPauseInitializer = memo(() => {
   return null;
 });
 
+const AnimationsInitializer = memo(() => {
+  const { initializeAnimations } = useAnimationsStore();
+
+  useEffect(() => {
+    initializeAnimations();
+  }, [initializeAnimations]);
+
+  return null;
+});
+
 export const StoreInitializer = memo(() => {
   return (
     <>
       <OnchainActionsInitializer />
       <OnchainActionsRevertInitializer />
       <SoundInitializer />
+      <AnimationsInitializer />
       <TutorialInitializer />
       <AchievementsInitializer />
       <BalanceInitializer />
