+++ frontend/app/stores/useSoundStore.ts
@@ -10,7 +10,6 @@ const SOUND_VOLUME_KEY = "sound_volume";
 const MUSIC_ENABLED_KEY = "music_enabled";
 const MUSIC_VOLUME_KEY = "music_volume";
 const HAPTICS_ENABLED_KEY = "haptics_enabled";
-const ANIMATIONS_LEVEL_KEY = "animations_level";
 const FIRST_LAUNCH_KEY = "has_launched_before";
 
 const musicAssets: { [key: string]: any } = {
@@ -194,13 +193,10 @@ class SoundPool {
   }
 }
 
-export type AnimationLevel = "full" | "reduced" | "off";
-
 interface SoundState {
   isSoundOn: boolean;
   isMusicOn: boolean;
   isHapticsOn: boolean;
-  animationLevel: AnimationLevel;
   soundEffectVolume: number;
   musicVolume: number;
   musicPlayer: AudioPlayer | null;
@@ -216,7 +212,6 @@ interface SoundState {
   toggleSound: () => void;
   toggleMusic: () => Promise<void>;
   toggleHaptics: () => void;
-  setAnimationLevel: (level: AnimationLevel) => void;
   setSoundEffectVolume: (volume: number) => void;
   setMusicVolume: (volume: number) => void;
   playSoundEffect: (soundType: string, pitchShift?: number) => Promise<void>;
@@ -233,7 +228,6 @@ export const useSoundStore = create<SoundState>((set, get) => ({
   isSoundOn: false,
   isMusicOn: false,
   isHapticsOn: true,
-  animationLevel: "full",
   soundEffectVolume: 1,
   musicVolume: 0.2,
   musicPlayer: null,
@@ -256,7 +250,6 @@ export const useSoundStore = create<SoundState>((set, get) => ({
       const soundEnabled = await AsyncStorage.getItem(SOUND_ENABLED_KEY);
       const musicEnabled = await AsyncStorage.getItem(MUSIC_ENABLED_KEY);
       const hapticsEnabled = await AsyncStorage.getItem(HAPTICS_ENABLED_KEY);
-      const animationLevel = await AsyncStorage.getItem(ANIMATIONS_LEVEL_KEY);
       const soundVolume = await AsyncStorage.getItem(SOUND_VOLUME_KEY);
       const musicVolume = await AsyncStorage.getItem(MUSIC_VOLUME_KEY);
       const hasLaunchedBefore = await AsyncStorage.getItem(FIRST_LAUNCH_KEY);
@@ -316,7 +309,6 @@ export const useSoundStore = create<SoundState>((set, get) => ({
         isSoundOn: soundEnabled === "true" || soundEnabled === null,
         isMusicOn: musicOn,
         isHapticsOn: hapticsEnabled === "true" || hapticsEnabled === null,
-        animationLevel: (animationLevel as AnimationLevel) || "full",
         soundEffectVolume: soundVolume ? parseFloat(soundVolume) : 1,
         musicVolume: volume,
         musicPlayer: musicPlayer,
@@ -334,7 +326,6 @@ export const useSoundStore = create<SoundState>((set, get) => ({
         isSoundOn: true,
         isMusicOn: false,
         isHapticsOn: true,
-        animationLevel: "full",
         soundEffectVolume: 1,
         musicVolume: 0.5,
         soundPool: new SoundPool(),
@@ -413,11 +404,6 @@ export const useSoundStore = create<SoundState>((set, get) => ({
       AsyncStorage.setItem(HAPTICS_ENABLED_KEY, newValue.toString());
       return { isHapticsOn: newValue };
     });
-  },
-
-  setAnimationLevel: (level: AnimationLevel) => {
-    set({ animationLevel: level });
-    AsyncStorage.setItem(ANIMATIONS_LEVEL_KEY, level);
   },
 
   setSoundEffectVolume: (volume) => {
@@ -890,13 +876,11 @@ export const useSound = () => {
     isSoundOn,
     isMusicOn,
     isHapticsOn,
-    animationLevel,
     soundEffectVolume,
     musicVolume,
     toggleSound,
     toggleMusic,
     toggleHaptics,
-    setAnimationLevel,
     setSoundEffectVolume,
     setMusicVolume,
     playSoundEffect,
@@ -909,13 +893,11 @@ export const useSound = () => {
     isSoundOn,
     isMusicOn,
     isHapticsOn,
-    animationLevel,
     soundEffectVolume,
     musicVolume,
     toggleSound,
     toggleMusic,
     toggleHaptics,
-    setAnimationLevel,
     setSoundEffectVolume,
     setMusicVolume,
     playSoundEffect,
